+++ src/lib/builtins/bytes.ts
@@ -1,0 +1,7 @@
+/**
+ * @module kryo/builtins/bytes
+ */
+
+import { BytesType } from "../types/bytes";
+
+export const $Bytes: BytesType = new BytesType({maxLength: Infinity});